rule-scribe-games/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
pydantic==2.8.1
google-generativeai==0.5.0
qdrant-client==1.8.2
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
